+++ modules/apps/layout/layout-api/bnd.bnd
@@ -4,12 +4,12 @@ Bundle-Version: 8.2.0
 Export-Package:\
 	com.liferay.layout.adaptive.media,\
 	com.liferay.layout.configuration,\
+	com.liferay.layout.crawler,\
 	com.liferay.layout.exception,\
 	com.liferay.layout.friendly.url,\
 	com.liferay.layout.list.retriever,\
 	com.liferay.layout.model,\
 	com.liferay.layout.responsive,\
-	com.liferay.layout.search.util,\
 	com.liferay.layout.security.permission.resource,\
 	com.liferay.layout.service,\
 	com.liferay.layout.service.persistence,\
